virttest.utils_package: Add package manager module
New pacakge manager module for handle package operation on host or
with remote session.

Use LocalPackageMgr which is subclass from software_manager in
avocado to decide the package manager and install or remove
package on local host.

If remote session is given, use the RemotePackageMgr class
to find out the package manager on remote for install or remove
packages.

Signed-off-by: Wayne Sun <gsun@redhat.com>