bootstrap: Remove custom logging handling
The `virttest.bootstrap` redefines the logging, which is not really
fortunate as it removes the default Avocado handlers. Let's remove this
custom logging handling and rely on Avocado for this. This brings one
change which I deliberately didn't address and that is a different
logging format. Originally it was:

    18:37:14 INFO | qemu test config helper

now it is only:

    qemu test config helper

which I prefer given that this is suppose to be human interface and when
one runs this inside CI they usually get timestamp from it.

Also the output is colored (on supported terminals)

Related to this the `virttest.postprocess_iozone` had to be adjusted to
not use custom handling and rather rely on `logging.basicConfig` when
executed as a script (which doesn't work due to relative import, but
let's wait for some user's request to fix it).

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>