unattended_install: Remove the hack to get ks from providers
This hack was added in a1449342bc649af28e057044e8b32e929c1d48da to allow
spice team to override the stock kickstart files, but it is actually a
backdoor which picks the first matching file from any provider, no
matter which provider is actually being used.

Let's disallow this behavior and ask people to submit their files to the
main repository.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>