avocado_vt: Fix the params handling
Avocado-vt supports "params" and "avocado_params". In recent Avocado the
"params" is a property therefor we can use super(VirtTest).params to get
to it's value, but on 36lts the "params" is variable stored during
__init__. This was workaround previously by set of hacks, this version
replaces it with storing the values in `self.__params` and
`self.__avocado_params` during `__init__` making the whole process
easier to follow (and hopefully with less bugs in it).

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>