Bootstrap: generate host os configuration
This introduces the generation of another file doing bootstrap time,
host-os.cfg.

This file contains the name, version and release of the host OS.  That
information comes, by default, from the Avocado distro detection, but
can be overriden by command line arguments.

This should allow for some tests in some providers, such as tp-qemu,
to properly enable or disable tests that should be available only
for some distros.

Signed-off-by: Cleber Rosa <crosa@redhat.com>