virttest: Support venv in data_dir
The data_dir only detects system-wide installation or from-git
execution. Let's use `sys.prefix` to check the current instance's
default path.

Currently most of the shared data are put in "/usr/share/..." for
system-wide and "." for virtual-env installations with the exception of
shared test-providers. I think it's a bug so lets treat all of
those-type files the same way.

Last but not least changing the virtual-env detection to recommended
solution as env variables are not really safe.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>