VT Bootstrap: reduce the number of fake command line arguments set
The args argument given to run is only used to be passed to
`bootstrap.boostrap()`, which only uses a very limited number of
options.

It feels safe enough to remove the fake ones that are not used.  And
even for fake ones used (`vt_config`, `vt_keep_image`) bootstrap
should actually use defaults if these are missing (more on that
later).

Signed-off-by: Cleber Rosa <crosa@redhat.com>