VT Bootstrap: do not use fake values to drive bootstrap
The `vt_config` and `vt_keep_image` "options" are hardcoded and never
set by the user, so it makes no sense to use them to drive the
download (and extraction) of the image.

The only currently valid option is `vt_no_downloads`, set by
`--vt-no-downloads`, so let's use that to determine if a download
*and* extraction is done.

If it's believed that another switch to control extraction *only* is
needed, then we can add something similar to `--vt-keep-image`.

Signed-off-by: Cleber Rosa <crosa@redhat.com>