Memory setting: avoid overwriting cartesian configuration value
We currently have the memory configuration set twice:

 1) etc/avocado.d/vt.conf
 2) shared/cfg/base.cfg

Configuration #1 is at the Avocado level, in a "ini" style file.
Configuration #2 is in Avocado-VT's own cartesian config.  Currently,
configuration #1 overwrites #2, no matter what is set in individual
sections of other cartesian config files.  So, if you design a test or
change a guest OS definition that needs more or less memory (and set
that in a cartesian based configuration item, "mem = xxxx") that will
be overwritten and not effective.

But, we can not set a fallback value in #1 that will apply to #2.  #1
will always overwrite #2, because it's applied last and takes the
highest precedence.

To overcome this, let's keep the #1 configuration empty, default to
None, and only apply it if set.  The behavior will still be such that
if a user sets a value in #1, it will overwrite whatever is in #2.  It
just won't happen *by default* because of the empty default value.

Signed-off-by: Cleber Rosa <crosa@redhat.com>