VT Bootstrap: do not consider the finished message another step
The last current step in the bootstrap process doesn't really do
anything.  It feels weird to me that we consider it another phase.

Let's replace that with a message saying that the bootstrap process
has finished, keeping the pointer to the Avocado-VT docs and adding
also a pointer to the Avocado docs.

Signed-off-by: Cleber Rosa <crosa@redhat.com>