VT Bootstrap: split copying of repo configuration and download phases
The first step, the copying of test providers repo configuration
(the ".ini" files) will always be done, from the local copy (usually
source code or files from a package) to the local datadir based
directory.

The second phase should be a download if the test providers have not been
downloaded yet (unless --vt-no-download is given) and then an update,
unless either --vt-update-providers *or* --vt-no-downloads is given.

This also changes the meaning of `--vt-no-downloads`, to mean that no
download (of any type) will be attempted.  The bootstrap process done
on Travis needed to be adjusted, since `--vt-no-downloads` should
really mean no downloads.

Signed-off-by: Cleber Rosa <crosa@redhat.com>