VT Bootstrap: introduce a new option to skip download of assets only
The command line options `--vt-no-downloads` had a misleading
implementation, it would still allow downloads (such as test providers).

That behavior was changed, and to replace the previous functionality,
that is, users that wants to skip just the download of asset files,
a new `--vt-skip-verify-download-assets` is being introduced.

We may think about adding similar options to the other phases of
the bootstrap process if it makes sense.

Signed-off-by: Cleber Rosa <crosa@redhat.com>