avocado_vt: Rework --no-download arguments in a backward compatible way
The behavior of `--no-download` argument was recently changed in a
non-backward compatible way, which require changes to many existing
scripts/jobs out there. Let's keep the old meaning the same, but also
introduce meaningful options.

This is not backward compatible with the changes introduced recently but
as this happened inside one sprint I think we should not keep the newly
introduced but unreleased option.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>