virttest.bootstrap: Decrease the spam at the end of bootstrap
The UPPER CASE message at the end really hits my eyes, one would assume
it's something super important which might break your system, but in the
end it's just an advert for Avocado-vt and also Avocado documentation.

I don't see a reason for such a big banner and also I don't think we
need to spam users with two links when the first link points to Avocado.
Besides, when users already executed `avocado vt-bootstrap`, they are
also able to run `avocado run`.

Last but not least, even this banner is a bit too much as people can
still use `virttest` as a library in combination with Autotest. Anyway
we don't get any responses whether it's true so let's just minimize the
banner and keep it there as it might be useful.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>