Fix HMP parsing of 'info block' on QEMU >= 1.5
'info block- can generate output such as:

drive_image1 (#block188): /path/to/image.qcow2 (qcow2)
    Cache mode:       writeback

When parsing those lines, the device name was including the extra
information (' (#block188)' in this example), which makes matches
against the information present on qtree fail.

This doesn't happen when a QMP monitor is used, though, so it only
affects tests running with HMP and with QEMU >= 1.5.0.

Signed-off-by: Cleber Rosa <crosa@redhat.com>