avocado_vt: Improve the nettype detection
Currently when nettype is not set in config it's set according to
provider. For qemu it's always "user". This commit changes the default
(only for qemu) to either "bridge" (when running as root) or "user"
(otherwise). This should give better user experience.

Note the default for spice or other providers hasn't changed, only
slight simplification was made.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>