Pass '-U' sub option to qemu-img to force shared permissions since qemu 2.10
qemu commit 244a566 introduces image locking to perm operations.
When we run avocado-vt to test a guest which image format is qcow2, If
host OFD lock is available and vt.setup.keep_guest_running is True or
option 'check_image' is yes, The 'qemu-img info' or the 'qemu-img check'
reports error due to it can't obtain the shared "write" lock on opened
images, This patch fixes the issue.

Signed-off-by: Lin Ma <lma@suse.com>