virttest.qemu_devices: Refresh AAVMF_VARS on force_create_image
The purpose of `force_create_image` is to start fresh, usually on
unattended_install tests. So far the AAVMF_VARS was only created once,
but the old state of this file might spoil the results. Let's re-create
it also when `force_create_image` is enabled on the main `image1` image.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>