scripts: Create logging_config lib
With the avocado logging refactors, the LoggingConfig
object ended up removed from the utils_misc library,
but there were a couple of scripts using them. As
those scripts haven't been used very often, and since
inspektor was ignoring legitimate syntax problems,
this passed and nobody noticed.

So let's reinstate the LoggingConfig class, but in a
separate lib only for the scripts. This makes the
scripts functional again and makes linting check to PASS.

Signed-off-by: Lucas Meneghel Rodrigues <lookkas@gmail.com>