virttest.qemu_vm: Do respect pci_bus on "activate_nic"
The "qemu_vm.activate_nic" hotplugs the nic device, but currently it
does not store, nor respect the "pci_bus". This makes it impossible to
define into which bus the nic should be placed and on some setups fails
as the default bus does not support hotplugging, even though there are
other available buses.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>