Pass hvc0 console to CentOS kernel params
This fixes the Anaconda installation log not being shown in the serial console.

Below are some snippets of the actual installation logs that are now shown.
They help in identifying which step is currently being executed by Anaconda and
debugging when the installlation fails (for example, it shows yum failures).

2017-12-14 16:56:17: Welcome to ESC[0;31mCentOS Linux 7 (AltArch)ESC[0m!
...
2017-12-14 16:56:27: ESC[1;23rStarting installer, one moment...ESC[1;24rESC[H
...
2017-12-14 16:56:34: ESC[1;23rESC[HESC[2Banaconda 21.48.22.121-1 for CentOS AltArch 7 started.ESC[1;24rESC[HESC[3;54H
...
2017-12-14 18:27:44: OS install is completed