virttest.qemu_vm: Update internal representation on (de)activate_nic
The (de)activate_nic methods create a new device and hotplug it into the
running machine, but it does not updates the internal representation
(qemu_devices). The problem is, that the internal representation is then
broken and inserting device into some buses might even fail or require
additional changes.

Ideally we should use the `qemu_devices` hotplug capability, but that
would require many changes to the existing virtnet/qemu_vm code so this
version simply finds the matching buses, attemps to insert the device
and on success uses the updated `bus` to plug the device in. The `addr`
is currently not in use by the existing code so I'm keeping the same
behavior in this version as well.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>