virttest.qemu_devices: Add PCIE with root ports representation bus
For some time we are using the main system PCIE bus directly, which is
not recommended and does not allow hotplug. This patch creates a new
QPCIEWithRootPorts bus, which represents a PCIE bus with optionally
added root_ports to some addresses.

By default it behaves the same as QPCIBus, but when one inserts a device
into port, which has root_port attached (by `add_root_port`), instead of
inserting the device into the PCIE bus it changes the `bus` property to
the root_port assigned to that address.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>