q35: Put most devices into root_ports
On q35 it's recommended to create a root_port per each slot in main
system pcie bus and for hotplug it's required. This patch basically does
that by using QPCIEWithRootPorts internal representation of the main
pcie bus. As we don't know in advanced how many devices will be needed
for hotplug, this version adds root-port per each slot (except of 0x0 -
mch; 0x1 - vga when std is used and 0x1f - ich9).

Note: When "-vga std" is not used the 0x1 port will end up without
root-port, but it's very unlikely this port will be used for hotplug as
some devices are usually inserted before the hotplug devices. Anyway in
the future we might want to detect and insert root-port when vga is not
std.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>