virttest.qemu_monitor: Ignore errors on unpickle
The qemu monitor tries to establish connection on __init__ but it's
possible the qemu process is hanged/broken and the connection can't be
established. In such situation Monitor raises exception, which should be
treated in code.

Anyway the situation is a bit different, when unpickling the code as
there is no (easy) way of ignoring such issue, so let's just say
`suppress_exceptions` is always `True` for unpickled objects.
Theoretically this should be only possible on env-cleanup.

This fixes a hang when `env_postprocess` crashes before killing all of
the created machines, which causes the env.load() to fail during
func_at_exit, which leads to unfinished processes and Avocado hang after
producing all the results.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>