virttest.qemu_vm: Allow to disable pci_controllers autosort
I'm not sure why the autosort is there, because people should insert
devices in the order they want them to be inserted, but this definitely
breaks "pci_devices" test, which has some scenarios where
higher-priority device is inserted in lower-priority device. When the
autosort reorders the inserts it breaks as the lower-priority device is
not yet inserted.

In this version I left the autosort enabled by default, but ideally we
should change it to disabled by default and unless there is a great
benefit eventually remove the whole autosort completely. Anyway I
understand that might break lots of cases.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>