avocado_vt: Log full pre/test and post exceptions
Currently when both test and post stages fail, the post section only
logs the exception details which is followed by full traceback of the
test. This is quite confusing and makes it harder to follow what
actually happened. This patch immediately logs the full traceback where
it happens and repeats the most significant (test, post) exception
details in test (as usual).

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>