avocado_vt: Change the default guest-os in vt-bootstrap
Currently the vt-bootstrap's default guest-os is the default guest-os,
which is JeOS.25. There is just a single variant of this asset, so it
works well, but now with the JeOS.27 multiple archs were introduced
resulting in all of them being queried.

Let's assume by default that the user only wants to check-out the
current host's architecture. Users can always use:

    --vt-guest-os JeOS.27

to get all of them (or JeOS.27.ppc64 to force-download ppc64 variant on
ppc64le machine).

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>