virttest.qemu_devices: Improve the cd_format fixtures
Some architectures don't support ide or the "None" cd_format and are
automatically translated to "scsi-cd" instead. The thing is that scsi-cd
on those archs usually require additional tweaks and it's only performed
when ide/None is used, but not when "scsi-cd" format is used.

This usually works as those additional tweaks are part of the
cartesian_config, but it fails when test specifies cdrom of this
particular type. Let's have the hook be performed on such occasions as
well.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>