utils_env: Support various ip sniffers
Make the framework support various ip sniffers, and add the support of
tshark (wireshark).

The reason to do so is because tcpdump till now (the current latest
version is 4.9.2) has some problem to deal with the bootp/dhcp packet:
it will not show the 'vendor extensions/options' field when a packet
has the 'sname' or 'file' field, and that would forbid `address_cache`
being updated properly. So let us use tshark to workaround this issue.

Signed-off-by: Xu Han <xuhan@redhat.com>