Set default value for param scsi_hba
When launching VM with scsi devices on some platforms, the following
error would be emitted if param scsi_hba is not set.

  DeviceError: Failed to insert device:
  ...
      driver = virtio-scsi-pci
      ...
  ParentBus({'aobject': 'pci.0'}): No matching bus

The patch fixes this error by setting the default value of param
scsi_hba to virtio-scsi-pci since virtio-scsi-pci does be the
default option of the images_define_by_variables function.

Signed-off-by: Xu Han <xuhan@redhat.com>