VirtTestLoader: Remove machine type for guest_listing and arch_listing
The guest name should be independent from machine type. It is not
necessary to list machine type for --vt-list-guests. The machine
type is not like arch, it actually does not matter whether it is
included in the guest name. However, it has to be filtered by
machine.cfg, so remove the suffix after the parser finish its work.

Signed-off-by: qizhu <qizhu@redhat.com>