env_process: Refresh "vm" on "vt_type" change
It's possible to use 2 different "vt_types" within a single job, but
"env_process" does not cope with that (it used to be impossible) and
simply proceeds with vm object of different class, when available in
env. This patch adds "isinstance(vm, expected_class)" check that makes
sure we use the right object and refreshes it otherwise.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>