ip_sniffing: Fix the issue of dropping cache
Under concurrency conditions, a `KeyError` would be raised since
the given `HWADDR` (key) record might have been already dropped
in the context of other thread.

We can fix this issue by checking if the given record is in cache
firstly, so that to ensure we will not drop a non-existent record.

Signed-off-by: Xu Han <xuhan@redhat.com>