avocado_vt: Improve "avocado_params" handling
For some time Avocado uses "self.params" property which points to
"self.__params". Avocado-vt used to override directly this property, but
it's better to simply create a new "self.__params_vt" property which
should allow the core Avocado to access the original params directly via
"self.__params".

In order to correctly support 52lts we still need to tweak the
"get_state", but once 52lts is deprecated we can remove that code.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>