qemu_vm: Console connection should be ready before sending 'cont'
The original design left a little time for consoles to be created,
and made consoles missed the head part of some logs, so advance the
consoles' creation in this commit. This scheme can work in my test,
if there is same issue next time, please find a more efficient way.

Signed-off-by: Sitong Liu <siliu@redhat.com>