Call "CmdResult.*_text" methods in backward compatible way
The fda9b8934 uses new "CmdResult.stdout_text" and
"CmdResult.stderr_text" methods, which are not part of 52.x LTS release.
Let's create a new "compat_52lts" file to store wrappers needed to run
52.X and master and use it to get stdout/stderr from process.CmdResult.

On some places pure "stdout" or "stderr" is used and bytes are used to
compare with. This should work always on master and ony 52.X it'd
compare string to bytes. This works well on python2 but fails on
python3, but python3 was not supported on 52.X so I believe it's fine.

Occasionally I used the raw "stdout" or "stderr" to write to binary
files. Again, on python2 it should work and on master we'll be writing
bytes.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>