migration inner function
Allow a function to be called after the start of migration.
The function is one of an internal list, and the parameter
is set e.g.

   [("postcopy", 70), ("somethingelse", "aparameter")]

with the second value being used as a parameter to the function.
Initially implment postcopy, where the parameter is an
upper end of a random bound on the 1st pass % to switch to postcopy on.

Signed-off-by: Dr. David Alan Gilbert <dgilbert@redhat.com>