shared: Add Fedora.27 for all archs
This commit removes the existing Fedora.27.aarch64 and replaces it with
generic Fedora.27 cfg for all archs, which should be easier to maintain.

It was tested thoroughly on x86_64; partially on ppc64, ppc64le and
s390x (started installation dvd/url) via tcg and it was not checked on
aarch64 (Avocado-vt failed to create make_create_command), but there is
just a single difference and that is "ks=cdrom:/ks.cfg" vs.
"ks=cdrom" but the later is already used in JeOS (which is basically the
same with a different ks).

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>