Use autotest libs on autotest control files
It makes no sense to depend on Avocado libraries on files that are
supposed to be run by Autotest, as it actually increases the
requirements and complexity of the code.

This is a partial revert of 7b7c5960ede3d2201775d1fd8f47c1df8d7d413c.

Reference: https://github.com/avocado-framework/avocado-vt/issues/1512
Signed-off-by: Cleber Rosa <crosa@redhat.com>