Undefine libvirt vm using param during postprocess
This patch introduces a new param 'kill_vm_libvirt' to undefine
a libvirt vm during test postprocess given that param is set.

This would be helpful incase of cleaning the env properly for
next test with right set of params(create_vm_libvirt) we can
create a newer vm based on given params for next test.

Signed-off-by: Satheesh Rajendran <sathnaga@linux.vnet.ibm.com>