avocado_vt: Fix monitor_type configuration
The configuration option for monitor_type does not really work for
several reasons.

1. It only sets type to qmp, which became the default long time ago
2. Even when used it overrides the "monitors", which breaks all tests
   that define additional monitors.

To address this issue I renamed the default monitor from "qmpmonitor1"
to just "monitor1" to avoid calling "hmp" monitor "qmpmonitor1" and
changed the overall "monitor_type" instead of trying to change that
default monitor (which might not exist in some tests).

The drawback of this solution is that some tests require certain monitor
type and such value would be overridden. This is why the default is not
set and therefor nothing is enforced unless one changes the
configuration.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>