shared: Add Fedora.28 for all archs
This adds Fedora.28 based on all archs. I tried only tcg installation
which failed on all non-x86 architectures usually crash or timeout, but
there was some progress. The x86 installs properly even though the "url"
variant (only) logs:

    Warning: unknown network kickstart URL: cdrom

This error is not logged in DVD installation, but even with the warning
our kickstart is used.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>