libvirt_vm: enablement of SAN based guest with LUN id
Boot guest with SAN based storage type using LUN id which can be
retrieved from /dev/disk/by-id, this enablement helps to use the
guest for SAN based migration tests. As LUN id would be same when
LUN is shared across multiple host and it is necessary for
migration to succeed.

Suggested-by: Satheesh Rajendran <sathnaga@linux.vnet.ibm.com>
Signed-off-by: Balamuruhan S <bala24@linux.vnet.ibm.com>