env_process: default to log libvirtd info
As the libvirtd is enabled by default to capture debug logs for
all the tests but archive it only for fail/error, it is better to
enable the log level to `info` so as to debug and get close to the
issue as well not polluting the archiving space.

Also back up of the libvirtd config should be done before enabling
and restarting the libvirtd is the right thing to do.

Suggested-by: Satheesh Rajendran <sathnaga@linux.vnet.ibm.com>
Signed-off-by: Balamuruhan S <bala24@linux.vnet.ibm.com>