Fix stress check command
Currently check_cmd would be assigned with stress_type, which will not hold
good all the time. Example, when we run netperf : stress_type would be netperf
but the actual command running on host would be netstress. This commit fixes
this by assigning check_cmd with actual stress command instead of stress_type.

Signed-off-by: Srikanth Aithal <sraithal@linux.vnet.ibm.com>