vm_xml:Add option 'debug' to method 'sync'
Lots of cases failed on 'sync' because the xml to be defined has some
kinds of error. We could not see the exact error message of the failed
'virsh define'. Add 'debug' as an option to enable us to recognize
problem without having to rerun the case or debug in other ways.