shared: Fix Fedora/RHEL unattended file ECHO command
The ECHO command is there to workaround issue where /proc/consoles fails
to list the right devices. Anyway as pointed in:

    https://bugzilla.redhat.com/show_bug.cgi?id=1351968#c7

it lists the right "major:minor", only the name is wrong. Let's switch
to using "major:minor" until this issue is resolved. (tested on x86,
ppc64le and s390x)

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>