utils_test: support different stress type on different VM
currently the stress type (stress or stress-ng or etc.,) is common across
all the VMs in multi vm environment. But to have different stress type
for different VM we do not have support. This patch enables it and the params
have to be suffixed with _VMNAME.

Example if we have VM1 name as virt-tests-vm1 and VM2 name as virt-tests-vm2
then the stress params should be like,

stress_args_virt-tests-vm1 = "--vm 2 --io 2 --cpu 2"
stress_args_virt-tests-vm2 = "-io 4 --cpu 3"

Signed-off-by: Balamuruhan S <bala24@linux.vnet.ibm.com>