virttest.qemu_devices: Fix arm's "-machine" workaround
Arm requires "-machine" to be always specified. We do have a workaround
for that and it works well with "execute_qemu", but it is not used when
initializing some of the variables, resulting in aarch64-pci not
supporting any devices at all. This patch detects the necessity earlier
on __init__ and then uses the "execute_qemu" to get the correct values.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>