Memory Hotplug Generic API's
The idea is to move commanly used memory hotplug functions
in libvirt_mem.py to virttest/utils_hotplug.py so that it
can be used by all.

As part of it have ported cream_mem_xml function.
Once this is accepted I will try to port all other fuctions
too.

Signed-off-by: Hariharan T.S <hari@linux.vnet.ibm.com>