shared: Remove check_zero_sym method from virtio_console_guest
This used to be used by removed "check_zero_sym" test, which used wrong
assumption that this symlink must always exists, but it doesn't have to
be there when other virtio device is present. Let's just rely only on
the "_get_port_status" test which makes sure the number of symlinks
correspondent to number of devices.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>