Add a possibility to keep running vms throughout jobs
In some cases (like reusing setup or state from a previous test
on a given vm), we would need to temporarily diregard the vm in
other tests possibly preparing other vms, all to be used together
in a final conjoint test. Analogously, we have to do the same for
the env file where the vms are registered.

Destroying the vms and the env file where they are registered
remains the default but now we also have the option to disregard
and thus keep them.

There are multiple immediate benefits from reusing previous vms
(and thus the env file) across jobs, here a few examples:

 - We can boot previously installed vms and do some manual customization
   on them, even develop a test more easily.

 - We can easily log into a vm and debug a test that failed and just
   completed but left the vm running in its final post-test state.

 - We can upload and download files to a vm, save it into a certain
   state (e.g. qemu snapshot), and reuse this state for a lot of tests
   without redoing it and thus saving a lot of time.

Signed-off-by: Plamen Dimitrov <pdimitrov@pevogam.com>