env_process: make the firewalld param with `default` option
`firewalld_service` param can hold `default` (leave it in
whatever state it is in), set `yes` to turn on and `no` to
turn off firewalld.

The recently introduced firewalld_service actually changes the setting
to work well on Ubuntu.18.10, but in reality it's a bug as firewalld
should not prevent that out of the box. Let's rename it to reflect the
fact of being a workaround and only enable it on that particular host
version (search for nftables firewalld libvirt to get more details).

Signed-off-by: Balamuruhan S <bala24@linux.vnet.ibm.com>
Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>