virttest: Rename the firewalld_service and disable it by default
The recently introduced firewalld_service actually changes the setting
to work well on Ubuntu.18.10, but in reality it's a bug as firewalld
should not prevent that out of the box. Let's rename it to reflect the
fact of being a workaround and only enable it on that particular host
version (search for nftables firewalld libvirt to get more details).

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>