utils_test: add check to update_boot_option() for add/remove actual args
currently update_boot_option() doesn't check whether the kernel parameters
passed as input is already existing and whether it is really required
to be added or not, the parameters can be in any order and when multiple
parameters are given as input. This patch ensures to check the existing
kernel parameters in /proc/cmdline and perform add/remove based on it.

Signed-off-by: Balamuruhan S <bala24@linux.vnet.ibm.com>