virttest.qemu_vm: Make sure all qemu processes finished
When forcefully destroying VM we only wait for the aexpect spawned
process (shell) to be destroyed but, especially on overloaded hosts, it
might happen that some processes are still running occupying write-lock.

This fixes a real issue where VM got destroyed but during postprocessing
"qemu-img info" failed as the left-behind process was still holding the
write lock to that image.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>